move Bytes to bits and start Tx(Bytes) subclass WIP
